+++ tests/pretalx/test_client.py
@@ -8,38 +8,38 @@ EVENT_SLUG = "pyconde-pydata-berlin-2023"
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_me_endpoint(pretalx_api):
-    me = pretalx_api.me()
+def test_me_endpoint(pretalx_client):
+    me = pretalx_client.me()
     assert me.name == "Florian Wilhelm"
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_events_endpoint(pretalx_api):
-    count, all_events = pretalx_api.events()
+def test_events_endpoint(pretalx_client):
+    count, all_events = pretalx_client.events()
     assert count == len(list(all_events))
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_event_endpoint(pretalx_api):
-    event = pretalx_api.event(EVENT_SLUG)
+def test_event_endpoint(pretalx_client):
+    event = pretalx_client.event(EVENT_SLUG)
     assert event.date_from == date.fromisoformat('2023-04-17')
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_submissions_endpoint(pretalx_api):
-    count, subs = pretalx_api.submissions(EVENT_SLUG)
+def test_submissions_endpoint(pretalx_client):
+    count, subs = pretalx_client.submissions(EVENT_SLUG)
     assert count == len(list(subs))
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_submission_endpoint(pretalx_api):
-    sub = pretalx_api.submission(EVENT_SLUG, 'MD9SLQ')
+def test_submission_endpoint(pretalx_client):
+    sub = pretalx_client.submission(EVENT_SLUG, 'MD9SLQ')
     assert sub.submission_type.en == 'Talk'
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_talks_endpoint(pretalx_api):
-    count, talks = pretalx_api.talks(EVENT_SLUG)
+def test_talks_endpoint(pretalx_client):
+    count, talks = pretalx_client.talks(EVENT_SLUG)
     assert count == len(list(talks))
 
 
@@ -47,8 +47,8 @@ def test_talks_endpoint(pretalx_api):
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_speakers_endpoint(pretalx_api):
-    count, speakers = pretalx_api.speakers(EVENT_SLUG)
+def test_speakers_endpoint(pretalx_client):
+    count, speakers = pretalx_client.speakers(EVENT_SLUG)
     assert count == len(list(speakers))
 
 
@@ -56,8 +56,8 @@ def test_speakers_endpoint(pretalx_api):
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_reviews_endpoint(pretalx_api):
-    count, reviews = pretalx_api.reviews(EVENT_SLUG)
+def test_reviews_endpoint(pretalx_client):
+    count, reviews = pretalx_client.reviews(EVENT_SLUG)
     assert count == len(list(reviews))
 
 
@@ -65,20 +65,20 @@ def test_reviews_endpoint(pretalx_api):
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_rooms_endpoint(pretalx_api):
-    count, rooms = pretalx_api.rooms(EVENT_SLUG)
+def test_rooms_endpoint(pretalx_client):
+    count, rooms = pretalx_client.rooms(EVENT_SLUG)
     assert count == len(list(rooms))
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_room_endpoint(pretalx_api):
-    room = pretalx_api.room(EVENT_SLUG, 1882)
+def test_room_endpoint(pretalx_client):
+    room = pretalx_client.room(EVENT_SLUG, 1882)
     assert room.name.en == 'B09'
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_questions_endpoint(pretalx_api):
-    count, questions = pretalx_api.questions(EVENT_SLUG)
+def test_questions_endpoint(pretalx_client):
+    count, questions = pretalx_client.questions(EVENT_SLUG)
     assert count == len(list(questions))
 
 
@@ -86,8 +86,8 @@ def test_questions_endpoint(pretalx_api):
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_answers_endpoint(pretalx_api):
-    count, answers = pretalx_api.answers(EVENT_SLUG)
+def test_answers_endpoint(pretalx_client):
+    count, answers = pretalx_client.answers(EVENT_SLUG)
     assert count == len(list(answers))
 
 
@@ -95,8 +95,8 @@ def test_answers_endpoint(pretalx_api):
 
 
 @pytest.mark.skipif(os.getenv('GITHUB'), reason="on Github")
-def test_tags_endpoint(pretalx_api):
-    count, tags = pretalx_api.tags(EVENT_SLUG)
+def test_tags_endpoint(pretalx_client):
+    count, tags = pretalx_client.tags(EVENT_SLUG)
     assert count == len(list(tags))